+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5.64 (Pending):
+* Add `ChannelMetaData` class that represents characteristics of each channel.
 * Add `ChangeUtil.toChanges` method that directly accepts `ChannelDescriptor`.
 * Move the conversion of sessions to json for rest service from ReplicantSession to EncoderUtil
   class in same package as AbstractSessionRestService.